Update README.md
This project explored the German Credit data set with the goal of developing a model that would help predict the creditworthiness of individuals applying for credit from the institution.  This was done by splitting the data into training and validation sets, creating a multiple regression model, then testing the model's ability to predict whether the person would default on their loan or not.